app/models/comment.py
"""
Comment 모델 — comments 테이블 (ERD Phase 1)

id         INTEGER PK
post_id    INTEGER FK→posts
user_id    INTEGER FK→users
content    TEXT
created_at DATETIME
"""

from sqlalchemy import Column, BigInteger, Text, DateTime, ForeignKey
from sqlalchemy.sql import func
from app.database import Base


class Comment(Base):
    __tablename__ = "comments"

    id = Column(BigInteger, primary_key=True)
    post_id = Column(BigInteger, ForeignKey("posts.id"), nullable=False)
    user_id = Column(BigInteger, ForeignKey("users.id"), nullable=False)
    content = Column(Text, nullable=False)
    created_at = Column(DateTime(timezone=True), server_default=func.now())
