requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
